+++ Plutarch/FFI.hs
@@ -35,6 +35,7 @@ import Plutarch.Builtin.Bool (PBool)
 import Plutarch.Builtin.ByteString (PByteString)
 import Plutarch.Builtin.Data (PData)
 import Plutarch.Builtin.Integer (PInteger)
+import Plutarch.Builtin.String (PString)
 import Plutarch.Builtin.Unit (PUnit)
 import Plutarch.Internal.Eq (PEq, (#==))
 import Plutarch.Internal.Generic (PCode)
@@ -50,6 +51,7 @@ import Plutarch.Internal.PlutusType (
  )
 import Plutarch.Internal.Quantification (PForall (PForall))
 import Plutarch.Internal.ScottEncoding (PlutusTypeScott)
+import Plutarch.Internal.Show (PShow)
 import Plutarch.Internal.Term (
   ClosedTerm,
   Config,
@@ -70,8 +72,6 @@ import Plutarch.Internal.Witness (witness)
 import Plutarch.List (PList, PListLike (PElemConstraint, pcons, pelimList, pnil), pconvertLists, plistEquals)
 import Plutarch.Maybe (PMaybe (PJust, PNothing))
 import Plutarch.Script (Script (Script))
-import Plutarch.Show (PShow)
-import Plutarch.String (PString)
 import PlutusTx.Builtins.Internal (BuiltinBool, BuiltinByteString, BuiltinData, BuiltinUnit)
 import PlutusTx.Code (CompiledCode, CompiledCodeIn (DeserializedCode), getPlc)
 import PlutusTx.Prelude (BuiltinString)
