+++ Plutarch/String.hs
@@ -7,26 +7,18 @@ module Plutarch.String (
   PString,
   -- Functions
   pisHexDigit,
-  pfromText,
   pencodeUtf8,
   pdecodeUtf8,
 ) where
 
-import Data.String (IsString, fromString)
-import Data.Text (Text)
-import Data.Text qualified as Text
-import GHC.Generics (Generic)
+import Plutarch.Builtin.String
+
 import Plutarch.Builtin.Bool (PBool, (#&&), (#||))
-import Plutarch.Builtin.ByteString (PByteString)
 import Plutarch.Builtin.Integer (PInteger)
-import Plutarch.Builtin.Opaque (POpaque)
 import Plutarch.Internal.Eq (PEq ((#==)))
-import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)
-import Plutarch.Internal.Newtype (PlutusTypeNewtype)
 import Plutarch.Internal.Numeric
 import Plutarch.Internal.Ord ((#<=))
 import Plutarch.Internal.PLam (plam)
-import Plutarch.Internal.PlutusType (DPTStrat, DerivePlutusType, PlutusType)
 import Plutarch.Internal.Term (
   S,
   Term,
@@ -34,47 +26,6 @@ import Plutarch.Internal.Term (
   (#),
   (:-->),
  )
-import Plutarch.Unsafe (punsafeBuiltin)
-import PlutusCore qualified as PLC
-
--- | Plutus 'BuiltinString' values
-newtype PString s = PString (Term s POpaque)
-  deriving stock (Generic)
-  deriving anyclass (PlutusType)
-
-instance DerivePlutusType PString where type DPTStrat _ = PlutusTypeNewtype
-
--- | @since WIP
-deriving via
-  (DeriveBuiltinPLiftable PString Text)
-  instance
-    PLiftable PString
-
-{-# DEPRECATED pfromText "Use `pconstant` instead." #-}
-
--- | Create a PString from 'Text'
-pfromText :: Text.Text -> Term s PString
-pfromText = pconstant
-
-instance IsString (Term s PString) where
-  fromString = pconstant . Text.pack
-
-instance PEq PString where
-  x #== y = punsafeBuiltin PLC.EqualsString # x # y
-
-instance Semigroup (Term s PString) where
-  x <> y = punsafeBuiltin PLC.AppendString # x # y
-
-instance Monoid (Term s PString) where
-  mempty = pconstant Text.empty
-
--- | Encode a 'PString' using UTF-8.
-pencodeUtf8 :: Term s (PString :--> PByteString)
-pencodeUtf8 = punsafeBuiltin PLC.EncodeUtf8
-
--- | Decode a 'PByteString' using UTF-8.
-pdecodeUtf8 :: Term s (PByteString :--> PString)
-pdecodeUtf8 = punsafeBuiltin PLC.DecodeUtf8
 
 {- | Verify if the given argument is the ASCII encoding of a hex digit. This
 includes specifically the following ASCII ranges (inclusively):
