+++ Plutarch/Trace.hs
@@ -23,6 +23,8 @@ module Plutarch.Trace (
 
 import Data.Kind (Type)
 import Plutarch.Builtin.Bool (PBool, pif)
+import Plutarch.Builtin.String (PString)
+import Plutarch.Internal.Show (PShow, pshow)
 import Plutarch.Internal.Term (
   Config (NoTracing, Tracing),
   LogLevel (LogDebug, LogInfo),
@@ -34,8 +36,6 @@ import Plutarch.Internal.Term (
   (#),
  )
 import Plutarch.Internal.Trace (ptrace, ptrace', ptraceDebug, ptraceInfo)
-import Plutarch.Show (PShow, pshow)
-import Plutarch.String (PString)
 
 {- | Like Haskell's @traceShowId@ but for Plutarch, at the info level.
